+++ Changelog.md
@@ -4,6 +4,7 @@
 ##### 2022-XX-YY
 
 - Added lower and upper annotation support to Invoices
+- Fixed `createInvoice::withClient` method when using with a Client model
 
 ## 1.0.1
 ##### 2022-10-11